Исправлены замечания наставника: 1.Добавил шаблон .env.sample 2.Переделал логику функции check_time_to_trigger_habit в habits.tasks. В функции теперь chat_id берется не из SETTINGS.py | .env files, при регистрации пользователя  учитывается данный параметр. Добавил источник chat_id к модели User, исправил тип данных на BigInteger. 3.в файл .flake8 добавил: exclude = migrations,settings.py, .venv Кроме того, исправил исключения, которые возникали при запуске flake8. Теперь код чист и по PEP8.

+++ users/models.py
@@ -5,13 +5,14 @@ from django.db import models
 
 NULLABLE = dict(null=True, blank=True)
 
+
 class User(AbstractUser):
     """Класс-пользователя"""
     email = models.EmailField(unique=True, verbose_name='Электронная почта', help_text='Введите электронную почту')
     phone = models.CharField(max_length=100, verbose_name='Телефон', **NULLABLE)
     city = models.CharField(max_length=100, verbose_name='Город', **NULLABLE)
     avatar = models.ImageField(upload_to='users/', verbose_name='Аватар', **NULLABLE)
-    chat_id = models.PositiveIntegerField(verbose_name='CHAT ID от телеграмма', **NULLABLE)
+    chat_id = models.BigIntegerField(verbose_name='CHAT ID от телеграмма', **NULLABLE)
 
     USERNAME_FIELD = 'username'
     REQUIRED_FIELDS = ['password', ]
